+++ ingeniamotion/__init__.py
@@ -1,7 +1,7 @@
 from . import enums
 from .motion_controller import MotionController
 
-__version__ = "0.8.0"
+__version__ = "0.8.1"
 """ str: Library version. """
 
 __all__ = ["MotionController", "enums"]