+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 ## [Unreleased]
 ### Added
-- Set the boot_in_app flag according to the file extension.
 - FSoE module.
 - STO example for Safe servo drives.
 
 ### Fix
 - Bug retrieving interface adapter name in Linux.
 - The stoppable_sleep method of the wizard tests does not block the main thread.
+
+## [0.8.1] - 2024-06-05
+### Added
+- Function to check if a configuration has been applied to the drive.
+- Set the boot_in_app flag according to the file extension.
 
 ## [0.8.0] - 2024-04-23
 ### Added